Move sqlRunner to service/
