+++ src/services/sqlRunner.ts
@@ -2,13 +2,13 @@ import * as Knex from 'knex';
 import * as path from 'path';
 import { reverse, keys } from 'ramda';
 
-import * as fs from './util/fs';
-import { dbLogger } from './logger';
-import Mapping from './domain/Mapping';
-import SqlCode from './domain/SqlCode';
-import * as promise from './util/promise';
-import SqlFileInfo from './domain/SqlFileInfo';
-import DatabaseObjectTypes from './enums/DatabaseObjectTypes';
+import * as fs from '../util/fs';
+import { dbLogger } from '../logger';
+import Mapping from '../domain/Mapping';
+import SqlCode from '../domain/SqlCode';
+import * as promise from '../util/promise';
+import SqlFileInfo from '../domain/SqlFileInfo';
+import DatabaseObjectTypes from '../enums/DatabaseObjectTypes';
 
 /**
  * SQL DROP statements mapping for different object types.
